Change interface to bytes to match beanstalkd
Beanstalkd does not explicitly enforce any encoding.  It's a byte protocol
whose commands happen to be readable as ASCII/UTF-8.  To match this behavior
in Python 3 without breaking bidirectional transparency, bytes should
be used instead of strings.

This change makes the use of bytes explicit, addressing earl/beanstalkc#60.

Note that bytes are usually inconvenient, so a convenience layer for strings
should be added.  This change merely makes the original Python 2
implementation's assumption of bytes explicit and correct.

This change obviates the need to deal with encoding.

Beanstalkd does implicitly encourage UTF-8 because it yields some data in
YAML.  PyYAML, which beanstalkc uses if available, additionally allows
UTF-16-LE and UTF-16-BE.  If a default encoding is chosen for the
convenience layer, it should probably be UTF-8.